refactor pages into proposals "app" containing "draft" sub-app
