--- web/src/grant-proposals/detail/page/local-draft-dialog-content.html
@@ -1,21 +1,0 @@
-<p>
-    You are viewing a temporary draft that has been saved in your browser's
-    local storage. Feel free to click the <strong>edit</strong> button on the
-    bottom of this page to continue editing, or <strong>save</strong> to
-    save the draft to be able to access it from another browser or computer.
-</p>
-<p>
-    Note that saving is optional, and that in order to <strong>save</strong>
-    you will be asked to sign in using either a traditional service (i.e., web
-    2.0, like Google, Github, etc.) or using the <a href="#">
-    polkadot{.js}</a> web3 extension. <strong>Either way, this will create an
-    account on our server with a traditional, centralised database, not the
-    Imbue Network blockchain</strong>. Saving would allow you to, for example,
-    share your draft before finalizing.
-</p>
-<p>
-    Only finalised proposals are eligible for funding. If you wish to go ahead
-    and <strong>finalise</strong> your proposal without creating an account,
-    your polkadot{.js} extension will ask you to cryptographically sign the
-    transaction to commit your proposal to the Imbue Network blockchain.
-</p>
