--- web/src/grant-submission/page/page.html
@@ -1,4 +1,0 @@
-<header>
-    <h1>Bring your project to life</h1>
-</header>
-<imbu-grant-submission-form id="form"></imbu-grant-submission-form>
