+++ web/src/proposals/detail/index.css
