+++ web/src/proposals/detail/index.html
