--- web/src/proposals/detail/page/page.css
@@ -1,172 +1,0 @@
-.top-wrapper {
-    margin-bottom: 10px;
-}
-#project-logo {
-    object-fit: cover;
-    height: 33vh;
-    width: 100%;
-}
-
-.project-name-heading {
-    width: 100%;
-}
-
-.action-buttons {
-    display: flex;
-    justify-content: space-between;
-}
-
-.action-button-container {
-    margin: 15px 0;
-    width: 32%;
-}
-
-.edit:not(:disabled) {
-    background-color: var(--secondary-button-background-color);
-    color: var(--secondary-button-color);
-    border: var(--secondary-button-border);
-}
-
-.edit:not(:disabled):hover {
-    background-color: var(--secondary-button-background-hover-color);
-}
-
-.button {
-    transition: all 200ms ease-in-out;
-    border-radius: 3px;
-}
-
-.button:disabled, .button:disabled:hover, .button.primary:disabled {
-    background-color: #676767;
-    color: var(--disabled-button-color);
-    cursor: unset;
-}
-
-.button.primary:not(:disabled) {
-    background-color: var(--mdc-theme-secondary);
-    color: white;
-}
-.button.primary:not(:disabled):hover {
-    background-color: var(--mdc-theme-primary);
-    /* text-shadow: 0px 0px 5px #000; */
-    color: black;
-}
-
-div {
-    color: var(--mdc-dialog-content-ink-color);
-}
-
-.mdc-tab .mdc-tab__text-label, .mdc-tab .mdc-tab__icon {
-    color: #fff;
-}
-
-header.project-name-header {
-    display: flex;
-    justify-content: space-between;
-}
-
-
-@media screen
-and (max-width: 380px) {
-    .action-buttons {
-        flex-direction: column;
-        width: 100% !important;
-    }
-    .action-button-container {
-        width: 100% !important;
-        margin: 0 0 10px 0;
-    }
-    .button {
-        /* padding: 8px; */
-        border-radius: 3px;
-    }
-}
-
-.tab-content {
-    margin: 20px 0;
-    min-height: 33vh;
-    display: none;
-}
-
-/* ol.tab-content {
-    list-style: none outside none;
-    padding-left: 0;
-}
-
-ol.tab-content li {
-    margin-bottom: 5px;
-} */
-
-#milestones li {
-    color: var(--mdc-theme-text-primary-on-background);
-    background-color: var(--mdc-theme-background);
-    margin: 5px 0;
-}
-
-#milestones .mdc-deprecated-list-item__secondary-text {
-    color: var(--mdc-theme-primary);
-}
-
-.tab-content.active {
-    display: block;
-}
-
-.tab-content a {
-    color: var(--mdc-theme-primary);
-}
-
-#funds-required {
-    color: #fff;
-    font-weight: bold;
-}
-
-ul.project-details {
-    color: #fff;
-    list-style: none outside none;
-    padding: 0;
-}
-ul.project-details li {
-    color: #fff;
-}
-
-#auth-dialog-web2-list li {
-    color: #fff;
-}
-
-
-
-#finalize.blob {
-	/* background: black; */
-	/* border-radius: 25%; */
-	/* margin: 10px; */
-	/* height: 20px; */
-	width: 32%;
-
-	box-shadow: 0 0 0 0 rgba(178, 255, 11, 1);
-	transform: scale(1);
-	animation: pulse 2s infinite;
-}
-
-#finalize.finalized {
-    width: 32%;
-    background-color: var(--theme-primary);
-    color: black;
-    font-weight: bolder;
-}
-
-@keyframes pulse {
-	0% {
-		transform: scale(0.95);
-		box-shadow: 0 0 0 0 rgba(178, 255, 11, 0.7);
-	}
-
-	70% {
-		transform: scale(1);
-		box-shadow: 0 0 0 10px rgba(255, 255, 255, 0);
-	}
-
-	100% {
-		transform: scale(0.95);
-		box-shadow: 0 0 0 0 rgba(178, 255, 11, 0.7);
-	}
-}
