+++ web/src/proposals/draft/editor/index.html
@@ -1,0 +1,4 @@
+<header>
+    <h1>Bring your project to life</h1>
+</header>
+<imbu-proposals-draft-submission-form id="form"></imbu-proposals-draft-submission-form>
