+++ web/src/proposals/draft/index.css
@@ -1,0 +1,3 @@
+.hidden {
+    display: none;
+}
