+++ web/src/proposals/draft/index.html
@@ -1,0 +1,8 @@
+<hq-pages id="pages" attr-for-selected="name" default-selection="not-found">
+    <imbu-proposals-draft-editor name="editor"></imbu-proposals-draft-editor>
+    <imbu-proposals-draft-preview name="preview"></imbu-proposals-draft-preview>
+    <!--
+        TODO: "imbu-proposals-draft-listing" i.e., "my drafts" where user can
+        choose from a list of drafts to edit/finalize, etc.
+    -->
+</hq-pages>
