+++ web/src/proposals/index.html
@@ -1,0 +1,5 @@
+<hq-pages id="pages" attr-for-selected="name" default-selection="not-found">
+    <imbu-proposals-draft name="draft"></imbu-proposals-draft>
+    <imbu-proposals-list name="listing"></imbu-proposals-list>
+    <imbu-proposals-detail name="detail"></imbu-proposals-detail>
+</hq-pages>
