+++ web/src/proposals/listing/index.css
@@ -1,0 +1,77 @@
+:host {
+    font-family: var(--theme-font-family);
+}
+
+h1 {
+    font-size: var(--theme-h1-size);
+    font-weight: var(--theme-h1-weight);
+}
+
+/*
+This is the flex version (attempt) at a CSS grid. The only reason to want to
+go this way, is that IE 11 doesn't support CSS grid, so we may want to figure
+out how to do some progressive enhancement, etc. to get this to work on IE.
+
+N.B., conditional comments don't work for IE 10 or 11.
+
+ol {
+    margin: 0;
+    padding: 0;
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: space-between;
+}
+ol::after {
+    content: "";
+    flex: auto;
+}
+
+imbu-proposal-item {
+    display: block;
+    width: 22%;
+    margin: 0 0 30px 0;
+    border: 1px solid #007;
+    box-shadow: 0 0 5px #333,
+                0 0 15px #333;
+}
+*/
+
+ol {
+    margin: 0;
+    padding: 0;
+    display: grid;
+    grid-template-columns: repeat(auto-fill, 23%);
+    /* grid-gap: 1%; */
+    justify-content: space-between;
+}
+
+imbu-proposal-item {
+    display: block;
+    margin: 0 0 30px 0;
+    border: 1px solid #007;
+    box-shadow: 0 0 5px #333,
+                0 0 15px #333;
+}
+
+@media screen and (max-width: 1100px) {
+    ol {
+        grid-template-columns: repeat(auto-fill, 31%);
+    }
+}
+
+@media screen and (max-width: 750px) {
+    ol {
+        grid-template-columns: repeat(auto-fill, 48%);
+    }
+}
+
+@media screen and (max-width: 500px) {
+    ol {
+        display: flex;
+        flex-flow: row wrap;
+        grid-template-columns: none;
+    }
+    imbu-proposal-item {
+        width: 100%;
+    }
+}
