+++ web/src/proposals/listing/index.html
@@ -1,0 +1,5 @@
+<h1>Discover Proposals</h1>
+
+<ol id="list">
+
+</ol>
