--- web/src/proposals/page/page.html
@@ -1,20 +1,0 @@
-<!-- FIXME: just use the generic dialog below instead of `imbu-auth-dialog` -->
-<imbu-auth-dialog id="auth-dialog"></imbu-auth-dialog>
-<hq-dialog id="dialog"></hq-dialog>
-
-<div class="top-wrapper">
-    <h2 class="section-title pro-tile">Discover Proposals</h2>
-</div>
-
-<div class="mdc-tab-bar" role="tablist" id="tab-bar">
-    <div class="mdc-tab-scroller">
-        <div class="mdc-tab-scroller__scroll-area">
-            <div role="list" class="w-dyn-items w-row">
-                <ol id="projects">
-                </ol>
-            </div>
-            <div class="mdc-tab-scroller__scroll-content">
-            </div>
-        </div>
-    </div>
-</div>
