+++ web/src/proposals/proposal-item/index.css
@@ -1,0 +1,37 @@
+:host {
+    display: block;
+}
+
+li {
+    list-style: none;
+}
+
+img {
+    object-fit: cover;
+    height: 200px;
+    width: 100%;
+}
+
+#name {
+    font-family: var(--theme-font-family);
+    font-size: 22px;
+    padding: 20px;
+    font-weight: 400;
+    line-height: 1.2;
+}
+
+a {
+    display: block;
+    text-decoration: none;
+    background: transparent;
+    transition: var(--base-transition);
+    padding: 20px 10px;
+    color: white;
+    font-size: 16px;
+    font-weight: 400;
+}
+
+a:hover {
+    background: var(--theme-primary);
+    color: #000;
+}
