+++ web/src/proposals/proposal-item/index.html
@@ -1,0 +1,6 @@
+<li>
+    <a id="contribute">
+    <img id="img">
+    <div id="name"></div>
+    </a>
+</li>
