+++ web/webpack.config.js
@@ -6,11 +6,8 @@ const WebpackBundleAnalyzer = require('webpack-bundle-analyzer');
 
 module.exports = {
     entry: {
-        "grant-submission/index": "./src/grant-submission/index.ts",
-        "proposals/index": "./src/proposals/index.ts",
-        "grant-proposals/detail": "./src/grant-proposals/detail.ts",
         "material-components": "./src/material-components.ts",
-        "dapp": "./src/dapp/dapp.ts",
+        "dapp": "./src/dapp/index.ts",
     },
     devtool: process.env.NODE_ENV === "development"
         ? "inline-source-map"
